Add service to get gallery images
Service and tests created to return all existing images in a gallery
for a specified subscription, resource group and gallery. The service
will be called by the WebAPI which in turn will be called by the
frontend and displayed to users, allowing them to create VMs from these
images or delete images.