Add service to get subscription resources
Service and tests created to return a subscription resource for a
specific subscription Id. The service will be called by the WebAPI
method to list all resource groups for a specific subscription, this is
done in the Azure SDK by first getting the subscription resource and
then getting all resource groups for the subscription resource.