Display Subscriptions in a table and add navigation to resource groups
MUI components have been added to the project and the subscription page
updated to display subscriptions in a MUI table. Subscription names can
now be clicked on to take users to the resource groups page.
Subscription page tests updated as weren't functioning correctly,
DataService was not mocked correctly and false posiitives were being
recorded. Refactored routing in the application so a route config can
be shared between the application and tests.